+++ LBEngine/Scene.cpp
@@ -113,7 +113,7 @@ void Scene::UpdateProjMatrix()
 	m_game.g_d3dDeviceContext->UpdateSubresource(m_game.g_d3dVSConstantBuffers[m_game.CB_Application], 0, nullptr, &proj, 0, 0);
 }
 
-void Scene::Update(DWORD delta)
+void Scene::Update(double delta)
 {
 	for (auto pSystem : m_nonDrawingSystems)
 	{
