+++ LBEngine/inc/CollisionSystem.h
@@ -5,7 +5,7 @@ class CollisionSystem :
 {
 public:
     CollisionSystem();
-    virtual void Execute(DWORD deltaTime) override;
+    virtual void Execute(double deltaTime) override;
 private:
     bool CheckCollision(Entity * pEntity1, Entity* pEntity2);
 };
